# tests: fixes for evtimer_msg and bench_runtime_coreapis