@fjmolinas 2020-03-24
The test failure seems unrelated:

```
--- run_test job results (1 failed, 20 passed, 21 total):
    failed:
    tests/unittests/esp32-wroom-32:gnu
```