@fjmolinas 2020-03-24
The test failure is still strange, maybe @gschorcht @kaspar030  wants to take a look, see if its just a random fail or a bug that got into master:

```
socat - open:/dev/ttyUSB0,b115200,echo=0,raw
READYs

START
...........................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................
rtt_rtc_tests.test_set_time (tests/unittests/tests-rtt_rtc/tests-rtt_rtc.c 46) exp 38 was 37
.
rtt_rtc_tests.test_set_alarm (tests/unittests/tests-rtt_rtc/tests-rtt_rtc.c 102) exp 38 was 37
.
rtt_rtc_tests.test_set_alarm_short (tests/unittests/tests-rtt_rtc/tests-rtt_rtc.c 138) exp 38 was 37
.
rtt_rtc_tests.test_set_alarm_set_time (tests/unittests/tests-rtt_rtc/tests-rtt_rtc.c 177) exp 38 was 37
..................................................................................................................................................................................................
run 976 failures 4
Timeout in expect script at "child.expect(r'OK \({} tests\)'.format(_tests), timeout=timeout)" (dist/pythonlibs/testrunner/__init__.py:52)
```